requirements.txt
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
certifi==2025.11.12
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageHash==4.3.2
iniconfig==2.3.0
numpy==2.4.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-telegram-bot==22.5
PyWavelets==1.9.0
scipy==1.16.3
SQLAlchemy==2.0.45
sqlmodel==0.0.27
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
